Make clippy happy(-er)
